+++ CHANGELOG.md
@@ -7,14 +7,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `getDisplayDetailsFromFormSubmissionPayment` to `paymentService`
+
 ## [7.2.0] - 2024-09-10
+
+### Added
+
+- `webhookSubscriptionService`
 
 ## [7.1.0] - 2024-09-03
 
 ### Added
 
 - `submissionService.processInjectablesInCustomResource()`
-- `webhookSubscriptionService`
 
 ## [7.0.0] - 2024-08-13